+++ src/bika/lims/content/calculation.py
@@ -71,6 +71,7 @@ schema = BikaSchema.copy() + Schema((
         'DependentServices',
         required=1,
         multiValued=1,
+        relationship="CalculationDependentServices",
         allowed_types=('AnalysisService',),
         widget=ReferenceWidget(
             checkbox_bound=0,