+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #2120 Support for Multiple component analysis
+- #2219 Make `UIDReferenceField` to not keep back-references by default
+- #2209 Migrate AnalysisRequest's ReferenceField fields to UIDReferenceField
+- #2218 Improve performance of legacy AT `UIDReferenceField`'s setter
 - #2214 Remove `DefaultContainerType` field (stale) from AnalysisRequest
 - #2215 Fix ParseError when search term contains parenthesis in widget search
 - #2213 Purge ComputedField fields from AnalysisRequest related with Profiles